코드업 기초 100제
--

+++ Algorithm/CodeUp/src/basic/basic_1012.java
@@ -7,7 +7,7 @@ public class basic_1012 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		
-		double a = sc.nextDouble();
+		float a = sc.nextFloat();
 		
 		System.out.println(a);
 	}
